HardwareSerialSetup: long -> unsigned long
because begin() requires it

+++ lib/DnWiFiDoorLock/DnWiFiDoorLock/App.h
@@ -55,7 +55,7 @@ namespace DnWiFiDoorLock {
 
             const char* const buildGitCommitHash;
 
-            const int serialBitsPerSecond;
+            const unsigned long serialBitsPerSecond;
 
             const uint8_t servoPin;
